requirements.txt
# python==3.6.12
# 音频基础处理
librosa==0.8.0
# 音频特征抽取
spafe==0.1.2
# 深度模型 docker image 已包含
# torch==1.3.1
# torchvision==0.4.2
# 数据计算
numpy==1.19.1
pandas==1.1.1
scikit-learn==0.23.2
matplotlib==3.3.0
tqdm==4.48.0
Flask_RESTful==0.3.8
Flask==1.1.1
gunicorn==20.0.4
requests==2.23.0
py_eureka_client==0.5.5
kafka-python==2.0.2
redis==3.3.11
numba==0.50.1